Adds lazyDefine
Adds bespoke polyfill specific `customElements.lazyDefine(name, classGenerator)` where `classGenerator` is a function that returns a custom elements class and is run when the polyfill first needs the class associated with `name`. This enables defining elements without incurring the cost of creating the element class or reading `observedAttributes` and the other callbacks. It can save non-trivial amounts of time on pages that load 100s of definitions but use only a small subset of them.