Adds `customElements.preferPerformance` setting and optimizes element walking code
Setting `customElements.preferPerformance` (1) will use querySelectorAll to find elements to upgrade, connect, and disconnect, (2) does not use the document construction observer to upgrade the document. This setting improves performance but reduces correctness in some cases: (1) removes support for upgrading custom elements inside HTMLImports, (2) does not upgrade undistributed custom elements in polyfilled shadow roots, and (3) elements do not upgrade with Mutation Observer timing, which is less spec correct.
Also introduces a number of minor optimizations to reduce the number of elements that need to be processed when upgrading, connecting, and disconnecting.