Adds `customElements.polyfillLazyDefine(name, constructorGenerator)`
This is an approximation of a proposed spec feature to allow definitions to be processed lazily when the first element to customize is found. The `constructorGenerator` is a function that's invoked only the first time and element is found to customize. It can return an element constructor function or a promise that resolves to an element constructor function.

Also refactored code between the CustomElementsRegistry and CustomElementsInternal to support this change.