Adds `polyfillLazyDefine`
Fixes #200 by adding customElements.polyfillDefineLazy(name, constructorGenerator) and adds a preferPerformance option.

The customElements.polyfillDefineLazy is an approximation of a potential spec feature to allow definitions to be processed lazily when the first element to customize is found. The constructorGenerator is a function that's invoked only the first time and element to customize is found. It can return an element constructor function or a promise that resolves to an element constructor function.